+++ sc/memiavl/iterator.go
@@ -2,9 +2,9 @@ package memiavl
 
 import (
 	"bytes"
-	dbm "github.com/tendermint/tm-db"
 
 	"github.com/sei-protocol/sei-db/common/utils"
+	dbm "github.com/tendermint/tm-db"
 )
 
 var _ dbm.Iterator = (*Iterator)(nil)
