+++ ss/pebbledb/bench_test.go
@@ -3,6 +3,7 @@ package pebbledb
 import (
 	"testing"
 
+	"github.com/sei-protocol/sei-db/config"
 	sstest "github.com/sei-protocol/sei-db/ss/test"
 	"github.com/sei-protocol/sei-db/ss/types"
 )
@@ -10,7 +11,7 @@ import (
 func BenchmarkDBBackend(b *testing.B) {
 	s := &sstest.StorageBenchSuite{
 		NewDB: func(dir string) (types.StateStore, error) {
-			return New(dir)
+			return New(dir, config.DefaultStateStoreConfig())
 		},
 		BenchBackendName: "PebbleDB",
 	}
