+++ ss/pebbledb/db_test.go
@@ -3,16 +3,16 @@ package pebbledb
 import (
 	"testing"
 
-	"github.com/stretchr/testify/suite"
-
+	"github.com/sei-protocol/sei-db/config"
 	sstest "github.com/sei-protocol/sei-db/ss/test"
 	"github.com/sei-protocol/sei-db/ss/types"
+	"github.com/stretchr/testify/suite"
 )
 
 func TestStorageTestSuite(t *testing.T) {
 	s := &sstest.StorageTestSuite{
 		NewDB: func(dir string) (types.StateStore, error) {
-			return New(dir)
+			return New(dir, config.DefaultStateStoreConfig())
 		},
 		EmptyBatchSize: 12,
 	}
