+++ ss/sqlite/bench_test.go
@@ -6,6 +6,7 @@ package sqlite
 import (
 	"testing"
 
+	"github.com/sei-protocol/sei-db/config"
 	sstest "github.com/sei-protocol/sei-db/ss/test"
 	"github.com/sei-protocol/sei-db/ss/types"
 )
@@ -13,7 +14,7 @@ import (
 func BenchmarkDBBackend(b *testing.B) {
 	s := &sstest.StorageBenchSuite{
 		NewDB: func(dir string) (types.StateStore, error) {
-			return New(dir)
+			return New(dir, config.DefaultStateStoreConfig())
 		},
 		BenchBackendName: "Sqlite",
 	}
