+++ ss/sqlite/db_test.go
@@ -6,6 +6,7 @@ package sqlite
 import (
 	"testing"
 
+	"github.com/sei-protocol/sei-db/config"
 	sstest "github.com/sei-protocol/sei-db/ss/test"
 	"github.com/sei-protocol/sei-db/ss/types"
 	"github.com/stretchr/testify/suite"
@@ -15,7 +16,7 @@ import (
 func TestStorageTestSuite(t *testing.T) {
 	s := &sstest.StorageTestSuite{
 		NewDB: func(dir string) (types.StateStore, error) {
-			return New(dir)
+			return New(dir, config.DefaultStateStoreConfig())
 		},
 		EmptyBatchSize: 0,
 	}
