+++ tools/cmd/benchmark/generate.go
@@ -1,4 +1,4 @@
-package main
+package benchmark
 
 import (
 	"fmt"
