+++ tools/cmd/operations/prune.go
@@ -1,9 +1,11 @@
-package main
+package operations
 
 import (
 	"fmt"
 
+	"github.com/sei-protocol/sei-db/config"
 	"github.com/sei-protocol/sei-db/ss"
+	"github.com/sei-protocol/sei-db/tools/cmd/benchmark"
 	"github.com/spf13/cobra"
 )
 
@@ -34,7 +36,7 @@ func prune(cmd *cobra.Command, _ []string) {
 		panic("Must provide db backend")
 	}
 
-	_, isAcceptedBackend := ValidDBBackends[dbBackend]
+	_, isAcceptedBackend := benchmark.ValidDBBackends[dbBackend]
 	if !isAcceptedBackend {
 		panic(fmt.Sprintf("Unsupported db backend: %s\n", dbBackend))
 	}
@@ -49,7 +51,9 @@ func prune(cmd *cobra.Command, _ []string) {
 // Prunes DB at given height
 func PruneDB(dbBackend string, dbDir string, version int64) {
 	// TODO: Defer Close Db
-	backend, err := ss.NewStateStoreDB(dbDir, dbBackend)
+	ssConfig := config.DefaultStateStoreConfig()
+	ssConfig.Backend = dbBackend
+	backend, err := ss.NewStateStore(dbDir, ssConfig)
 	if err != nil {
 		panic(err)
 	}
